Apps/FrontendSoftware/src/BIL/requirements.txt
pyinstaller
protobuf~=4.21.6
grpcio~=1.51.1
ulid-py
### N.B. If using a Mac and having connection authentications issues (such as an
### error indicating "[SSL: CERTIFICATE_VERIFY_FAILED]"), this is possibly due
### to Apple using outdated SSL certificates. Python can be made to use more
### updated ones from certifi by executing the script found at:
### `/Applications/Python <WHATEVER VERSION>/Install Certificates.command`
###
pymongo >= 3.8.0
dnspython
pyzmq
Pillow
numpy
scipy
certifi